Adds TypeDeclar, SumCon cases to simplify
I'm not sure what should be updated in the case of a Sum. Can the
simplified pipeline update the sumTName for example? How does that fit
into the update?

If the sumTDef is Nothing - what should the resulting job be? I've tried
to construct a noOpJob that doesn't change anything, but I don't know if
this is correct.