Berlin: Implements simplify for SumCon
One detail missed in the last commit is that all of the components of a
Context.Definition need to be processed as part of the job that's
returned by simplify.

For example the Context.Def has a defTerm and a defMTy that must both be
handled.

Only Def, TypeDeclar and SumCon need to be handled by simplify. simplify
is no-op for the other constructors of Context.Definition.