Berlin: Adds a pass which updates the currentNamespace
If the inPackageTrans pass is passed an `:in-package` Sexp then it will
update the namespace in the PassArgument using the name specified in the
Sexp.

This pass is intended to be run using `registerAfterEachStep` to keep
the context namespace in-sync with the Sexps that are being processed in
the pipeline.