Adds initContext to initalize the Context from input
initContext is a factor of Contextify.op that takes all the input Sexps
and adds them to the context. The output of initContext will be used to
initialize the Berlin pipeline.

In its current form no Tracing or Feedback is possible because it still
uses the effect mechanism from Contextify. The functions that
initContext calls need to be updated to use the effects in the Env
module to integrate with the Berlin pipeline