Berlin: Add relocated form to job
A pass on the infoTable can promote a form from the table to the
context. In this case it will return a Pipeline.InContext form. We keep
a reference to this form in the table using a `Relocated` sexp.

This commit fixes the intoTable pass processing by actually adding the
Pipeline.InContext form to the newForms field of the job.