Read the docs (#1056)
* Start docs revisions

* Remove no-longer-accurate sections

* Remove old stuff from core docs

* Remove old stuff from erased core docs

* Update current type rules to new judgement style etc

* Add code documentation

* Hide LLVM docs for now

* Enable llvm docs

* Better haddock display

* Update compiler diagram

* Update compiler overview

* Use requirements.txt in documentation readme

At the moment the installation instructions say to install all of the
dependencies separately (also, `rtd-tikz` is missing). Instead, we
should say to just use the dependency list.

* Rename the local server target to `serve`

The install instructions say to run `make html; make latexpdf`, but the Makefile overwrites sphinx's `html` target with the local server stuff.

* More core documentation updates (#1070)

* Fixes in primitive description

* Remove some TODOs

(Σ/Unit aren't in the version with records that this doc is going to
describe)

* Note about variable binding

* Replace some colons with \Chk

* Add `let` bindings

* Use a different symbol for usage compat and subtype

* Stack input Γ/Θ

It seems like bussproofs cannot put premises on multiple lines,
so i have to try and save space

* Add notes about what is an in/output in judgements

* Add macros like \FunTy etc

* Add coproducts

* Remove bussproofs and abstract some other macros

bussproofs doesn't allow multiple lines of premises, has verbose markup,
and has a hard limit of 5 premises (which we might exceed in complex
rules). I am actually not sure what its positives are, but they don't
seem to apply to this document.

* add primitive type arity

* Add record type rules and β rule

* Fix typo

* Some more description about records

* Replace \Rightarrow with \mapsto in case

Since \Rightarrow/\Leftarrow are used for syn/chk

* Use case-style syntax for unpacking records

* Add some more description about records/coproducts

* Add some brackets, tweak some spacing

* Update parsing docs

* Fix reduction rules in docs for bidi syntax

As written previously, most of the β rules replaced a variable
with a *term*, not an *elimination*, which is invalid syntax.
The corrected rules are a bit noisier with the new annotations, but
they're just shuffling around data that is already immediately
available.

* Add rule for type annotations

Also remove the usage from annotations (see #1041), since it is always
known by the typechecker from the surrounding context (unlike the type,
which might not be inferrable), and otherwise the reduction rules (which
*don't* know it) would have to guess.

* Doc: removed excess white space, refactor examples to current Juvix 1.0 syntax

* Doc: Update syntax guide to be more accurate

* Update golden files

* Remove commented out code

* Update golden tests

* [ci] changes from CI

Co-authored-by: Christopher Goes <cwgoes@pluranimity.org>
Co-authored-by: Andy Morris <andy@heliax.dev>
Co-authored-by: mariari <mariari@protonmail.com>
Co-authored-by: drone <noreply@drone>