Propogate errors from passes into the Berlin pipeline
Berlin pipeline errors are captured using the Feedback.Eff effect. This
commit adds this effect to the constraints of Passes and updates the
throwSexp method to add errors to Feedback.

The Errors from passes are made Sexp serializable so they can be stored
in Feedback messages

Finally the errors from context passes are connected with the Berlin
pipeline in the initContext function.