Contextify: Check for symbols locally in atomResolution first.
Check in the local module for a symbol before looking up in the
qualified symbol table.

Consider the case where a module B contains the symbol `fi` and the
symbol `fi` is also available in the local module.

We want `main` to use the `fi` from the local module instead of the one
available in B. So we need to check the local module first.

```
mod D where

open B

sig fi : 3
let fi = 3

sig main : 3
let main = fi
```