Initial revision of Categorial term
Introduce a module which implements syntax-checking,
typechecking, evaluation, erasure, and code generation
for a new modular term which can be combined with existing
Core terms in the style of data types à la carte.
(That is, Core terms do not need to know the structure of
Categorial terms, and Categorial terms do not need to know
the structure of Core terms.  They exchange information
via monad-transformer interfaces.)

Some of its sub-modules are `Private` (by convention, since
Haskell does not explicitly have the concept), meaning that
they can share internal definitions which are not intended
to be exposed outside the `Categorial` module.