Add a new "Interpreter" backend
This is another backend -- a peer of LLVM, Plonk, and
Michelson, which implements its own `Parameterisation` --
which "compiles" to a Haskell ADT and "executes" it
through reduction by Haskell code.  Its purposes are to
allow faster development of new term types (since it's
likely to be easier to interpret them in Haskell than in,
say, LLVM) and to help with fault isolation (since one
can begin a new term type by testing with Interpreter and
getting it working that way before implementing LLVM code
generation for the new term type).